+++ examples/upgradable-example/scripts/deploy-box-beacon.ts
@@ -26,7 +26,7 @@ async function main() {
 
     const chainId = await hre.network.provider.send('eth_chainId', []);
     if (chainId === '0x12c') {
-        let _ = hre.run('verify:verify', { address: await box.getAddress() });
+        const _ = hre.run('verify:verify', { address: await box.getAddress() });
     }
 }