+++ packages/hardhat-zksync-chai-matchers/test/fixture-projects/hardhat-project/hardhat.config.ts
@@ -18,6 +18,7 @@ const config: HardhatUserConfig = {
             url: 'http://0.0.0.0:3050',
             ethNetwork: 'ethNetwork',
             zksync: true,
+            forceDeploy: true,
         },
     },
 };