+++ packages/hardhat-zksync-solc/src/constants.ts
@@ -10,6 +10,8 @@ export const DETECT_MISSING_LIBRARY_MODE_COMPILER_VERSION = '1.3.14';
 export const USER_AGENT =
     'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/120.0.0.0 Safari/537.36';
 
+export const ZKSOLC_COMPILER_PATH_VERSION = 'local_or_remote';
+
 export const defaultZkSolcConfig: ZkSolcConfig = {
     version: 'latest',
     compilerSource: 'binary',
@@ -62,7 +64,7 @@ export const COMPILERS_CONFLICT_ZKVM_SOLC = (version: string) =>
 export const MISSING_LIBRARIES_NOTICE =
     'zksolc compiler detected missing libraries! For more details, visit: https://era.zksync.io/docs/tools/hardhat/compiling-libraries.html.';
 export const COMPILE_AND_DEPLOY_LIBRARIES_INSTRUCTIONS =
-    'To compile and deploy libraries, please run: `yarn hardhat deploy-zksync:libraries --private-key <PRIVATE_KEY>`';
+    'To compile and deploy libraries, please run: `yarn hardhat deploy-zksync:libraries`';
 export const MISSING_LIBRARY_LINK =
     'For more details on how to use deploy-zksync:libraries task from hardhat-zksync-deploy plugin, visit: https://era.zksync.io/docs/tools/hardhat/hardhat-zksync-deploy.html.';