+++ packages/hardhat-zksync-solc/src/index.ts
@@ -41,6 +41,7 @@ import {
     COMPILE_AND_DEPLOY_LIBRARIES_INSTRUCTIONS,
     MISSING_LIBRARY_LINK,
     COMPILING_INFO_MESSAGE_ZKVM_SOLC,
+    ZKSOLC_COMPILER_PATH_VERSION,
 } from './constants';
 import { ZksolcCompilerDownloader } from './compile/downloader';
 import { ZkVmSolcCompilerDownloader } from './compile/zkvm-solc-downloader';
@@ -64,6 +65,7 @@ const zkVmSolcCompilerDownloaderMutex = new Mutex();
 const zkSolcCompilerDownloaderMutex = new Mutex();
 
 extendConfig((config, userConfig) => {
+    defaultZkSolcConfig.version = userConfig.zksolc?.settings?.compilerPath ? ZKSOLC_COMPILER_PATH_VERSION : 'latest';
     config.zksolc = { ...defaultZkSolcConfig, ...userConfig?.zksolc };
     config.zksolc.settings = { ...defaultZkSolcConfig.settings, ...userConfig?.zksolc?.settings };
     config.zksolc.settings.optimizer = {
@@ -150,6 +152,7 @@ subtask(TASK_COMPILE_SOLIDITY_GET_COMPILATION_JOBS, async (args, hre, runSuper) 
         if (!isZksolcDownloaded) {
             await zksolcDownloader.downloadCompiler();
         }
+
         hre.config.zksolc.settings.compilerPath = zksolcDownloader.getCompilerPath();
         hre.config.zksolc.version = zksolcDownloader.getVersion();
     });