+++ packages/hardhat-zksync-viem/src/internal/clients.ts
@@ -6,13 +6,14 @@ import {
 import { EthereumProvider } from "hardhat/types";
 import { Address, Chain, WalletClientConfig } from "viem";
 import { Eip712WalletActions, eip712WalletActions } from "viem/zksync";
+import { WalletClient } from "./types";
 
 export { getPublicClient, innerGetPublicClient, getTestClient };
 
 export async function getWalletClients(
   provider: EthereumProvider,
   walletClientConfig?: Partial<WalletClientConfig>
-): Promise<Eip712WalletActions[]> {
+): Promise<(WalletClient &Eip712WalletActions)[]> {
   const { getAccounts } = await import("./accounts");
   const { getChain } = await import("./chains");
   const chain = walletClientConfig?.chain ?? (await getChain(provider));
@@ -24,7 +25,7 @@ export async function getWalletClient(
   provider: EthereumProvider,
   address: Address,
   walletClientConfig?: Partial<WalletClientConfig>
-): Promise<Eip712WalletActions> {
+): Promise<WalletClient & Eip712WalletActions> {
   const { getChain } = await import("./chains");
   const chain = walletClientConfig?.chain ?? (await getChain(provider));
   return (
@@ -37,7 +38,7 @@ export async function innerGetWalletClients(
   chain: Chain,
   accounts: Address[],
   walletClientConfig?: Partial<WalletClientConfig>
-): Promise<Eip712WalletActions[]> {
+): Promise<(WalletClient & Eip712WalletActions)[]> {
   const viem = await import("viem");
   const { isDevelopmentNetwork } = await import("./chains");
   const defaultParameters = isDevelopmentNetwork(chain.id)