+++ packages/hardhat-zksync-viem/test/fixture-projects/import/contracts/Import.sol
@@ -3,7 +3,6 @@
 pragma solidity ^0.8.0;
 pragma abicoder v2;
 
-// import Foo.sol from current directory
 import "./Bar.sol";
 
 contract Import {
@@ -12,9 +11,6 @@ contract Import {
     constructor(){
 
     }
-    // Initialize Bar
-
-    // Test Foo.sol by getting it's name.
     function getFooName() public view returns (string memory) {
         return bar.foo().name();
     }