+++ packages/hardhat-zksync-viem/test/fixture-projects/library/contracts/Lib.sol
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: MIT
+pragma solidity ^0.8.11;
+library Lib {
+    function plus(uint a, uint b) public pure returns (uint) {
+        return a - b;
+    }
+}