+++ packages/hardhat-zksync-vyper/src/index.ts
@@ -14,7 +14,7 @@ import { Mutex } from 'hardhat/internal/vendor/await-semaphore';
 import './type-extensions';
 import chalk from 'chalk';
 import { CompilationJob } from 'hardhat/types';
-import { ZkArtifacts } from './artifacts';
+import { PROXY_NAME, ProxtContractOutput, ZkArtifacts, proxyNames } from './artifacts';
 import { compile } from './compile';
 import { checkSupportedVyperVersions, pluralize } from './utils';
 import {
@@ -22,10 +22,14 @@ import {
     defaultZkVyperConfig,
     TASK_COMPILE_VYPER_CHECK_ERRORS,
     TASK_COMPILE_VYPER_LOG_COMPILATION_ERRORS,
+    ZKVYPER_COMPILER_PATH_VERSION,
 } from './constants';
 import { ZkVyperCompilerDownloader } from './compile/downloader';
 
 extendConfig((config, userConfig) => {
+    defaultZkVyperConfig.version = userConfig.zkvyper?.settings?.compilerPath
+        ? ZKVYPER_COMPILER_PATH_VERSION
+        : 'latest';
     config.zkvyper = { ...defaultZkVyperConfig, ...userConfig?.zkvyper };
     config.zkvyper.settings = { ...defaultZkVyperConfig.settings, ...userConfig?.zkvyper?.settings };
     config.zkvyper.settings.optimizer = {
@@ -78,10 +82,17 @@ subtask(TASK_COMPILE_VYPER_RUN_BINARY, async (args: { inputPaths: string[]; vype
     delete compilerOutput.zk_version;
     delete compilerOutput.long_version;
 
-    if (compilerOutput.__VYPER_FORWARDER_CONTRACT) {
-        (hre.artifacts as ZkArtifacts).forwarderOutput = compilerOutput.__VYPER_FORWARDER_CONTRACT;
-        delete compilerOutput.__VYPER_FORWARDER_CONTRACT;
-    }
+    proxyNames.forEach((proxyName) => {
+        if (compilerOutput[proxyName]) {
+            const proxyContractOutput: ProxtContractOutput = {
+                proxyName: proxyName as PROXY_NAME,
+                output: compilerOutput[proxyName],
+            };
+
+            (hre.artifacts as ZkArtifacts).proxyContractOutput = proxyContractOutput;
+            delete compilerOutput[proxyName];
+        }
+    });
 
     (hre.artifacts as ZkArtifacts).compilerOutput = compilerOutput;